Use local copy as source for build
This reduces build time greatly, and also makes local development much
easier, as no needs to Dockerfile are required. User can use Docker as
local development environment, which removes the need to install
dependencies on a local machine.

This also makes it lot easier to track dependencies, as everything must
be found in Docker image for local environment to work.